+++ src/Castle.Windsor/MicroKernel/IKernelInternal.cs
@@ -51,7 +51,7 @@ namespace Castle.MicroKernel
 
 		IDisposable OptimizeDependencyResolution();
 
-		object Resolve(Type service, IDictionary arguments, IReleasePolicy policy, bool ignoreParentContext = false);
+		object Resolve(Type service, Arguments arguments, IReleasePolicy policy, bool ignoreParentContext = false);
 
 		/// <summary>
 		/// Returns a component instance by the key
@@ -61,9 +61,9 @@ namespace Castle.MicroKernel
 		/// <param name = "arguments"> </param>
 		/// <param name = "policy"> </param>
 		/// <returns> </returns>
-		object Resolve(String key, Type service, IDictionary arguments, IReleasePolicy policy);
+		object Resolve(string key, Type service, Arguments arguments, IReleasePolicy policy);
 
-		Array ResolveAll(Type service, IDictionary arguments, IReleasePolicy policy);
+		Array ResolveAll(Type service, Arguments arguments, IReleasePolicy policy);
 
 		IHandler CreateHandler(ComponentModel model);